Fixed some issues and updating stuff
🛠️ Fixed an issue with the commandHandler:
In some cases the -execute command would error because of an oversight.
🛠️ Updated 2 of the commandBin commands to the current parameter layout
🛠️Updated the setupDB file as it wasn't updated to the current version

+++ commandHandler.js
@@ -13,7 +13,8 @@ function notCommandChannel(client, msg){
 		return false;
 	}
 
-	msg.delete();
+	if(!msg.deleted)
+		msg.delete();
 
 	return !(client.adminList.includes(msg.author.id) && client.s.ADMIN_COMMANDS);
 }
@@ -44,8 +45,10 @@ module.exports = {
 
         if(msg.content.startsWith(prefix)){
 
-			if(notCommandChannel(client, msg))
+			if(notCommandChannel(client, msg)){
 				return;
+			}
+
 
 			var tempMessage = msg.content.slice(prefix.length); // Remove prefix from string
 			var args = tempMessage.split(" ").filter(arg => arg !== '');
